Kealanba/prv gpu update

+++ src/gpu/intel/jit/gemm/generator/pieces/matrix_access.cxx
@@ -518,7 +518,7 @@ void BLASKernelGenerator<hw>::loadLoadStoreDescriptors(bool load, bool store, Re
             exdescStore.parts.extMessageLen = 0;
             descLoad.parts.responseLen = 0;
 
-            int underlyingSIMD = std::max<int>(block.simdSize, maxScatteredSIMD(hw, astrategy) >> 1);
+            uint32_t underlyingSIMD = std::max<uint32_t>(block.simdSize, (uint32_t)maxScatteredSIMD(hw, astrategy) >> 1);
             int log2GRFs = ilog2(underlyingSIMD * block.ebytes) - GRF::log2Bytes(hw);
             int log2Components = int(block.splitComplex);
